refactor(conference): move conferenceJoinUserRow do dto package
